Add single-image experiment with homeostatic rate control
- Add single-image experiment for testing image reconstruction through PING
- Implement homeostatic controller to maintain constant E rate across g_ei sweep
  - tune_I_E_for_target_rate() iteratively adjusts drive to hit target rate
  - hotloop_burnin() for fast rate measurement during tuning
- Add new analysis modules: base_metrics, calculate_regime_label, energy_metrics
- Update metrics experiment with g_ei sweep
- Reorganize phase-dependence experiment (underscore to hyphen)
- Add coupled-networks experiment stub
- Update content and artifacts

+++ pinglab/src/pinglab/analysis/pairwise_spike_count_corr.py
@@ -50,7 +50,7 @@ def pairwise_spike_count_corr(
     # Only consider spikes within [0, T)
     mask = (times >= 0.0) & (times < T)
     times = times[mask]
-    ids = ids[mask]
+    ids = ids[mask].astype(int)
 
     # Bin indices for each spike
     bin_idx = (times / bin_ms).astype(int)
